+++ data/src/main/java/com/puzzlebench/cmk/data/repository/CharacterDataRepository.kt
@@ -1,12 +1,12 @@
 package com.puzzlebench.cmk.data.repository
 
 import com.puzzlebench.cmk.data.mapper.repository.CharacterMapperRepository
-import com.puzzlebench.cmk.data.repository.source.ICharacterDataSource
+import com.puzzlebench.cmk.data.repository.source.CharacterDataSource
 import com.puzzlebench.cmk.domain.model.Character
 import com.puzzlebench.cmk.domain.repository.CharacterRepository
 
 
-class CharacterDataRepository constructor(private val dataSource: ICharacterDataSource,
+class CharacterDataRepository constructor(private val dataSource: CharacterDataSource,
                                           private val mapper: CharacterMapperRepository) : CharacterRepository {
 
     override fun save(c: List<Character>) {
